+++ packages/typespec-python/scripts/eng/regenerate.ts
@@ -29,10 +29,6 @@ const AZURE_EMITTER_OPTIONS: Record<string, Record<string, string> | Record<stri
     "azure/client-generator-core/usage": {
         namespace: "specs.azure.clientgenerator.core.usage",
     },
-    "azure/core/lro/rpc": {
-        "package-name": "azurecore-lro-rpc",
-        "namespace": "azurecore.lro.rpc",
-    },
     "client/structure/default": {
         namespace: "client.structure.service",
     },
@@ -182,10 +178,6 @@ const EMITTER_OPTIONS: Record<string, Record<string, string> | Record<string, st
     "type/union": {
         "package-name": "typetest-union",
         "namespace": "typetest.union",
-    },
-    "azure/core/lro/rpc": {
-        "package-name": "azurecore-lro-rpc",
-        "namespace": "azurecore.lro.rpc",
     },
     "client/structure/multi-client": {
         "package-name": "client-structure-multiclient",